Auth and User endpoint

+++ src/main/java/com/apr/car_sales/dtos/bid/BidDto.java
@@ -1,6 +1,5 @@
 package com.apr.car_sales.dtos.bid;
 
-import com.apr.car_sales.dtos.car.CarDto;
 import com.apr.car_sales.dtos.user.UserDto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class BidDto {
-    private int id;
+    private long id;
     private double bidAmount;
     private double askPrice;
     private UserDto bidder;
